Add game logic to switch between player x and o
- Add new game button to index.html. This is a temporary measure to
initialize an empty game board and turn counter of 0. The project
API has a functionality to initialize an empty game board.
The new game button does not have an associated UI for now.

- Add an object - boxIdAssignment to translate tic-tac-toe box ID from
string to numbers to game-engine.js

- Add a function - whichPiece to return either x pieces, or o pieces
depending on the turn (i.e. player) to game-engine.js

- Add a function - updateBoard to update the javascript game grid with
the appropriate game piece in the desired location to game-engine.js

- Add a function - onUserMove to events.js that triggers updateBoard
with the user's chosen box ID as parameter

- Add a function - gameFailure to ui.js to alert user that they haven't
initialized a game, and can't make a move.

- Add event handler in app.js to app.js to trigger onUserMove when
user clicks a tic-tac-toe board they want to add a game piece to